# common-docs/openapi/openapi.v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Aini-Inu API",
        "description": "Aini-Inu \ubc31\uc5d4\ub4dc REST API \ubb38\uc11c",
        "contact": {
            "name": "Aini-Inu Backend Team"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "\uac8c\uc2dc\uae00 \ubaa9\ub85d \uc870\ud68c",
                "description": "\ubb34\ud55c \uc2a4\ud06c\ub864 \uae30\ubc18\uc73c\ub85c \uac8c\uc2dc\uae00 \ubaa9\ub85d\uc744 \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "operationId": "getPosts",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseSliceResponsePostResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Community"
                ],
                "summary": "\uac8c\uc2dc\uae00 \uc0dd\uc131",
                "description": "\uc0c8 \uac8c\uc2dc\uae00\uc744 \uc0dd\uc131\ud569\ub2c8\ub2e4.",
                "operationId": "create_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePostResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/posts/{postId}/like": {
            "post": {
                "tags": [
                    "Community"
                ],
                "summary": "\uc88b\uc544\uc694 \ud1a0\uae00",
                "description": "\uac8c\uc2dc\uae00 \uc88b\uc544\uc694 \uc0c1\ud0dc\ub97c \ud1a0\uae00\ud569\ub2c8\ub2e4.",
                "operationId": "toggleLike",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePostLikeResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/posts/{postId}/comments": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "\ub313\uae00 \ubaa9\ub85d \uc870\ud68c",
                "description": "\uac8c\uc2dc\uae00 \ub313\uae00 \ubaa9\ub85d\uc744 \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "operationId": "getComments",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseSliceResponseCommentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Community"
                ],
                "summary": "\ub313\uae00 \uc791\uc131",
                "description": "\uac8c\uc2dc\uae00\uc5d0 \ub313\uae00\uc744 \uc791\uc131\ud569\ub2c8\ub2e4.",
                "operationId": "createComment",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseCommentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/posts/{postId}": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "\uac8c\uc2dc\uae00 \uc0c1\uc138 \uc870\ud68c",
                "description": "\uac8c\uc2dc\uae00 \ub2e8\uac74 \uc0c1\uc138\ub97c \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "operationId": "getPost",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePostDetailResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Community"
                ],
                "summary": "\uac8c\uc2dc\uae00 \uc0ad\uc81c",
                "description": "\uac8c\uc2dc\uae00 \uc791\uc131\uc790\uac00 \uac8c\uc2dc\uae00\uc744 \uc0ad\uc81c\ud569\ub2c8\ub2e4.",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Community"
                ],
                "summary": "\uac8c\uc2dc\uae00 \uc218\uc815",
                "description": "\uac8c\uc2dc\uae00 \uc791\uc131\uc790\uac00 \ub0b4\uc6a9\uc744 \uc218\uc815\ud569\ub2c8\ub2e4.",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePostResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/posts/{postId}/comments/{commentId}": {
            "delete": {
                "tags": [
                    "Community"
                ],
                "summary": "\ub313\uae00 \uc0ad\uc81c",
                "description": "\ub313\uae00 \uc791\uc131\uc790 \ub610\ub294 \uac8c\uc2dc\uae00 \uc791\uc131\uc790\uac00 \ub313\uae00\uc744 \uc0ad\uc81c\ud569\ub2c8\ub2e4.",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResponseVoid": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "type": "object",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "PostCreateRequest": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 0
                    },
                    "caption": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 0
                    },
                    "imageUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "maxItems": 5,
                        "minItems": 0
                    },
                    "contentProvided": {
                        "type": "boolean"
                    },
                    "resolvedContent": {
                        "type": "string"
                    }
                }
            },
            "ApiResponsePostResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/PostResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "PostResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    },
                    "content": {
                        "type": "string"
                    },
                    "imageUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "likeCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "commentCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "liked": {
                        "type": "boolean"
                    }
                }
            },
            "ApiResponsePostLikeResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/PostLikeResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "CommentCreateRequest": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 1
                    }
                },
                "required": [
                    "content"
                ]
            },
            "ApiResponseCommentResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/CommentResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "PostUpdateRequest": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 0
                    },
                    "caption": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 0
                    },
                    "imageUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "maxItems": 5,
                        "minItems": 0
                    },
                    "contentProvided": {
                        "type": "boolean"
                    },
                    "resolvedContent": {
                        "type": "string"
                    }
                }
            },
...
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "ApiResponseSliceResponsePostResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/SliceResponsePostResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
            "SliceResponsePostResponse": {
                "type": "object",
                "description": "\ubb34\ud55c \uc2a4\ud06c\ub864\uc6a9 Slice \uc751\ub2f5",
                "properties": {
                    "content": {
                        "type": "array",
                        "description": "\uc870\ud68c\ub41c \ud56d\ubaa9 \ubaa9\ub85d",
                        "items": {
                            "$ref": "#/components/schemas/PostResponse"
                        }
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32",
                        "description": "\ud604\uc7ac \ud398\uc774\uc9c0 \ubc88\ud638 (0-base)",
                        "example": 0
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32",
                        "description": "\ud398\uc774\uc9c0 \ud06c\uae30",
                        "example": 20
                    },
                    "first": {
                        "type": "boolean",
                        "description": "\uccab \ud398\uc774\uc9c0 \uc5ec\ubd80",
                        "example": true
                    },
                    "last": {
                        "type": "boolean",
                        "description": "\ub9c8\uc9c0\ub9c9 \ud398\uc774\uc9c0 \uc5ec\ubd80",
                        "example": false
                    },
                    "hasNext": {
                        "type": "boolean",
                        "description": "\ub2e4\uc74c \ud398\uc774\uc9c0 \uc874\uc7ac \uc5ec\ubd80",
                        "example": true
                    }
                }
            },
            "ApiResponsePostDetailResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/PostDetailResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "ApiResponseSliceResponseCommentResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/SliceResponseCommentResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}